requirements.txt
gradio>=4.0.0
rembg
Pillow
onnxruntime